# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "This is the API documentation for the project.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://hooshba.webcomcoai.com/api",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/regions": {
            "get": {
                "tags": [
                    "Regions"
                ],
                "summary": "دریافت لیست محله‌ها",
                "description": "این متد لیست محله‌ها را با قابلیت فیلتر بر اساس وضعیت انتشار بازمی‌گرداند.",
                "operationId": "getRegions",
                "parameters": [
                    {
                        "name": "type",
                        "in": "query",
                        "description": "فیلتر وضعیت انتشار محله‌ها",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "published"
                            ],
                            "example": "published"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "لیست محله‌ها با موفقیت دریافت شد.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "description": "لیست محله‌ها",
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "links": {
                                            "description": "لینک‌های صفحه‌بندی",
                                            "type": "object"
                                        },
                                        "meta": {
                                            "description": "اطلاعات صفحه‌بندی",
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "دریافت لیست محله‌ها با شکست مواجه شد.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "دریافت محله ها با شکست مواجه شد."
                                        },
                                        "message": {
                                            "description": "پیام خطا",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Regions"
                ],
                "summary": "ایجاد محله جدید",
                "description": "این متد برای ایجاد یک محله جدید استفاده می‌شود.",
                "operationId": "createRegion",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "نام محله",
                                        "type": "string",
                                        "example": "منطقه 1"
                                    },
                                    "slug": {
                                        "description": "نامک محله (اختیاری)",
                                        "type": "string",
                                        "example": "manategh-1"
                                    },
                                    "ordering": {
                                        "description": "ترتیب نمایش (اختیاری)",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "is_published": {
                                        "description": "وضعیت انتشار محله (اختیاری)",
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "محله با موفقیت ایجاد شد.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "محله ها با موفقیت ایجاد شد."
                                        },
                                        "province": {
                                            "description": "اطلاعات محله ایجاد شده",
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "اعتبارسنجی شکست خورد.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "اعتبارسنجی شکست خورد."
                                        },
                                        "details": {
                                            "description": "جزئیات خطاهای اعتبارسنجی",
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "ایجاد محله با شکست مواجه شد.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "ایجاد محله ها با شکست مواجه شد."
                                        },
                                        "message": {
                                            "description": "پیام خطا",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
...
        "/regions/front-single-region": {
            "get": {
                "tags": [
                    "Front Regions"
                ],
                "summary": "دریافت اطلاعات یک محله منتشر شده",
                "description": "این متد جزئیات مربوط به یک محله منتشر شده را به همراه استان‌های مرتبط نمایش می‌دهد.",
                "operationId": "frontSingleRegion",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "query",
                        "description": "اسلاگ محله‌ای که باید دریافت شود",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "اطلاعات محله با موفقیت دریافت شد.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "منطقه 1"
                                        },
                                        "provinces": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 101
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "استان تهران"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "دریافت محله منتشر شده با شکست مواجه شد.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "دریافت محله منتشر شده با شکست مواجه شد."
                                        },
                                        "message": {
                                            "description": "پیام خطا",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        }
    },
...
}